[REFACTOR] class diagram에 따라 패키지 위치 리팩토링
