--- src/main/java/org/hanihome/hanihomebe/member/controller/MemberController.java
@@ -1,4 +1,0 @@
-package org.hanihome.hanihomebe.member.controller;
-
-public class MemberController {
-}
