+++ src/main/java/org/hanihome/hanihomebe/member/web/MemberController.java
@@ -1,0 +1,4 @@
+package org.hanihome.hanihomebe.member.web;
+
+public class MemberController {
+}
