+++ src/main/java/org/hanihome/hanihomebe/security/auth/application/filter/SecurityHeadersFilter.java
@@ -1,4 +1,4 @@
-package org.hanihome.hanihomebe.auth.filter;
+package org.hanihome.hanihomebe.security.auth.application.filter;
 
 
 import jakarta.servlet.Filter;
